old/requirements_lmstudio.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
python-multipart==0.0.6
chromadb==0.4.18
sentence-transformers==2.2.2
torch==2.2.0 --index-url https://download.pytorch.org/whl/cpu
openai==1.3.9
pypdf==3.17.0
python-docx==1.1.0
aiofiles==23.2.1
python-dotenv==1.0.0
loguru==0.7.2
psutil==5.9.6
transformers==4.35.2
accelerate==0.24.1
